stocksprediction/requirements.txt
Django==2.2.6
django-phone-field==1.8.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
pytz==2019.3
sqlparse==0.3.0
